scaffolding refactor
