+++ src/main/java/jsges/nails/DTO/TipoServicioDTO.java
@@ -1,0 +1,8 @@
+package jsges.nails.DTO.servicios;
+
+import jsges.nails.DTO.TipoObjetoDTO;
+
+public class TipoServicioDTO extends TipoObjetoDTO {
+
+
+}
