--- src/main/java/jsges/nails/DTO/servicios/TipoServicioDTO.java
@@ -1,8 +1,0 @@
-package jsges.nails.DTO.servicios;
-
-import jsges.nails.DTO.TipoObjetoDTO;
-
-public class TipoServicioDTO extends TipoObjetoDTO {
-
-
-}
