+++ src/main/java/jsges/nails/domain/Cliente.java
@@ -1,0 +1,43 @@
+package jsges.nails.domain.organizacion;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+
+
+import java.io.Serializable;
+import java.util.Date;
+
+@Entity
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@ToString
+public class Cliente implements Serializable {
+        @Id
+        @GeneratedValue(strategy = GenerationType.IDENTITY)
+        private Integer id;
+
+        @Column(columnDefinition = "TEXT")
+
+        String razonSocial;
+        int estado;
+
+        @Column(columnDefinition = "TEXT")
+        String letra;
+
+        @Column(columnDefinition = "TEXT")
+        String contacto;
+
+        @Column(columnDefinition = "TEXT")
+        String celular;
+        @Column(columnDefinition = "TEXT")
+        String mail;
+
+        Date fechaInicio;
+        Date fechaNacimiento;
+
+
+}
