+++ src/main/java/jsges/nails/domain/TipoServicio.java
@@ -1,0 +1,17 @@
+package jsges.nails.domain.servicios;
+
+import jakarta.persistence.*;
+import jsges.nails.domain.TipoObjeto;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+
+@Entity
+@Data
+@ToString
+public class TipoServicio extends TipoObjeto {
+
+
+
+}
